+++ README.md
@@ -27,7 +27,7 @@
 - [ ] 회원을 삭제한다.
 
 ### 회원 정보 조회 기능
-- [ ] 회원 정보를 반환한다.
+- [x] 회원 정보를 반환한다.
 
 ### 패스워드 변경 기능
 - [ ] 이전 패스워드와 새로운 패스워드를 입력받아 새로운 패스워드로 변경한다.
