+++ src/main/java/woowacourse/shoppingcart/application/CustomerService.java
@@ -2,7 +2,9 @@ package woowacourse.shoppingcart.application;
 
 import org.springframework.stereotype.Service;
 import woowacourse.shoppingcart.dao.CustomerDao;
+import woowacourse.shoppingcart.domain.Customer;
 import woowacourse.shoppingcart.dto.CustomerCreateRequest;
+import woowacourse.shoppingcart.dto.CustomerResponse;
 
 @Service
 public class CustomerService {
@@ -22,5 +24,11 @@ public class CustomerService {
         if (customerDao.existsByNickname(nickname)) {
             throw new IllegalArgumentException("이미 존재하는 닉네임입니다.");
         }
+    }
+
+    public CustomerResponse findCustomerByEmail(String email) {
+        Customer customer = customerDao.findIdByEmail(email)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 회원입니다."));
+        return CustomerResponse.from(customer);
     }
 }
